requirements.txt
python-telegram-bot==20.7
openai==1.30.1
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2
psycopg2-binary==2.9.7